libexfat: avoid overflow in bitmap size calculation
BITS_PER used to contain a sizeof expression, which promoted the
cluster count calculation to size_t. After BITS_PER was changed to
__BITS_PER_LONG, DIV_ROUND_UP could be evaluated as 32-bit unsigned
arithmetic and wrap around for cluster counts near the exFAT limit.

Promote the cluster count to uint64_t before rounding and add a
regression test covering the maximum valid cluster count.

+++ include/libexfat.h
@@ -180,7 +180,7 @@ typedef __u32		bitmap_t;
 #define BITMAP_WORD_AT(bmap, bit)	((bitmap_t *)(bmap))[BIT_ENTRY(bit)]
 
 #define EXFAT_BITMAP_SIZE(__c_count)	\
-	(DIV_ROUND_UP(__c_count, BITS_PER) * sizeof(bitmap_t))
+	(DIV_ROUND_UP((uint64_t)__c_count, BITS_PER) * sizeof(bitmap_t))
 
 #define BITMAP_GET(bmap, bit)	\
 	(BITMAP_WORD_AT(bmap, bit) & BIT_MASK(bit))
